Remove session storage and move auth to common
